@@ -1,34 +1,34 @@
 <Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net9.0</TargetFramework>
-    <RootNamespace>QuantConnect.DataLibrary.Tests</RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="..\Demonstration.cs" Link="Demonstration.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="protobuf-net" Version="3.1.33" />
-    <PackageReference Include="NUnit" Version="4.2.2" />
-    <PackageReference Include="NUnit3TestAdapter" Version="4.6.0">
-      <PrivateAssets>all</PrivateAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.9.4" />
-    <PackageReference Include="Microsoft.TestPlatform.ObjectModel" Version="16.9.4" />
-    <PackageReference Include="QuantConnect.Algorithm" Version="2.5.*" />
-    <PackageReference Include="CsvHelper" Version="30.0.1" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
-    <PackageReference Include="NodaTime" Version="3.1.9" />
-  </ItemGroup>
-  <ItemGroup>
-    <Using Include="NUnit.Framework.Legacy.ClassicAssert" Alias="Assert" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\QuantConnect.DataBento\QuantConnect.DataSource.DataBento.csproj" />
-    <ProjectReference Include="..\..\Lean\Tests\QuantConnect.Tests.csproj" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Update="config.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
+	<PropertyGroup>
+		<Configuration Condition=" '$(Configuration)' == '' ">Release</Configuration>
+		<Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
+		<TargetFramework>net10.0</TargetFramework>
+		<IsPackable>false</IsPackable>
+		<TestProjectType>UnitTest</TestProjectType>
+		<OutputPath>bin\$(Configuration)\</OutputPath>
+		<Product>QuantConnect.Lean.DataSource.DataBento.Tests</Product>
+		<AssemblyName>QuantConnect.Lean.DataSource.DataBento.Tests</AssemblyName>
+		<RootNamespace>QuantConnect.Lean.DataSource.DataBento.Tests</RootNamespace>
+		<AssemblyTitle>QuantConnect.Lean.DataSource.DataBento.Tests</AssemblyTitle>
+		<AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
+	</PropertyGroup>
+	<ItemGroup>
+		<PackageReference Include="NUnit" Version="4.2.2" />
+		<PackageReference Include="NUnit.Analyzers" Version="3.6.1" />
+		<PackageReference Include="NUnit3TestAdapter" Version="4.6.0">
+			<PrivateAssets>all</PrivateAssets>
+		</PackageReference>
+		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.9.4" />
+	</ItemGroup>
+	<ItemGroup>
+		<Using Include="NUnit.Framework.Legacy.ClassicAssert" Alias="Assert" />
+	</ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\QuantConnect.DataBento\QuantConnect.DataSource.DataBento.csproj" />
+	</ItemGroup>
+	<ItemGroup>
+		<None Include="config.json">
+			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+		</None>
+	</ItemGroup>
 </Project>
