@@ -1,9 +1,8 @@
-
 <Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
 		<Configuration Condition=" '$(Configuration)' == '' ">Release</Configuration>
 		<Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-		<TargetFramework>net9.0</TargetFramework>
+		<TargetFramework>net10.0</TargetFramework>
 		<Product>QuantConnect.Lean.DataSource.DataBento</Product>
 		<AssemblyName>QuantConnect.Lean.DataSource.DataBento</AssemblyName>
 		<RootNamespace>QuantConnect.Lean.DataSource.DataBento</RootNamespace>
@@ -29,14 +28,14 @@
 		<OutputPath>bin\Release\</OutputPath>
 	</PropertyGroup>
 
-	<!--
+
 	<ItemGroup>
 		<Compile Include="..\..\Lean\Common\Properties\SharedAssemblyInfo.cs" Link="Properties\SharedAssemblyInfo.cs" />
 	</ItemGroup>
-	-->
+
 
 	<ItemGroup>
 		<PackageReference Include="QuantConnect.Lean.Engine" Version="2.5.*" />
-		<PackageReference Include="QuantConnect.Common" Version="2.5.*" />
 	</ItemGroup>
+
 </Project>
